requirements.txt
pyapriori~=0.1.0
cupy~=10.0.0
numpy~=1.20.3
parameterized~=0.8.1
scipy~=1.7.3
setuptools~=60.2.0
